+++ src/main/kotlin/lotto/domain/model/BonusNumber.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain.model
 
 class BonusNumber(private val bonusNumber: Int) {
 }
