+++ src/main/kotlin/lotto/domain/model/PurchaseAmount.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain.model
 
 class PurchaseAmount(private val purchaseAmount: Int) {
 }
